修改bug

+++ app/src/main/java/com/anxin/kitchen/bean/OrderInfoBean.java
@@ -44,12 +44,12 @@ public class OrderInfoBean {
         private int makeType;
         private int num;
         private int payType;
-        private int money;
-        private int mealFee;
+        private double money;
+        private double mealFee;
         private int tablewareId;
         private int tablewareType;
-        private int tablewareFee;
-        private int deliveryFee;
+        private double tablewareFee;
+        private double deliveryFee;
         private int deposit;
         private String contactPhone;
         private String contactName;
@@ -129,14 +129,14 @@ public class OrderInfoBean {
         public void setMoney(int money) {
             this.money = money;
         }
-        public int getMoney() {
+        public double getMoney() {
             return money;
         }
 
-        public void setMealFee(int mealFee) {
+        public void setMealFee(double mealFee) {
             this.mealFee = mealFee;
         }
-        public int getMealFee() {
+        public double getMealFee() {
             return mealFee;
         }
 
@@ -154,17 +154,17 @@ public class OrderInfoBean {
             return tablewareType;
         }
 
-        public void setTablewareFee(int tablewareFee) {
+        public void setTablewareFee(double tablewareFee) {
             this.tablewareFee = tablewareFee;
         }
-        public int getTablewareFee() {
+        public double getTablewareFee() {
             return tablewareFee;
         }
 
-        public void setDeliveryFee(int deliveryFee) {
+        public void setDeliveryFee(double deliveryFee) {
             this.deliveryFee = deliveryFee;
         }
-        public int getDeliveryFee() {
+        public double getDeliveryFee() {
             return deliveryFee;
         }
